LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/image-loading-lazy-move-into-script-disabled-iframe-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'http://localhost:8800/html/semantics/embedded-content/the-img-element/resources/subframe.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'http://localhost:8800/html/semantics/embedded-content/the-img-element/resources/subframe.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.


Harness Error (TIMEOUT), message = null

TIMEOUT A loading='lazy' image starts loading when the element is moved into
       an iframe where script is disabled Test timed out
